Improve input streaming
Previous attempts to split the input into Event and CommandResult
streams were clumsy, imprecise, and required a task.

A better approach is to just implement two Stream objects with
shared state in between them. Outputs which are not of the type
sought are placed onto a queue. That queue is read first---before
trying to read from the stream.