Allow combined stream/sink in ArqState
The ArqState contains helper functions which write data. These
require access to the incoming event streams as well, so we can
determine when the write finishes.

Previously, this was implemented as two objects: a sink and a
stream. Unfortunately, this does not match the implementation
in the AsyncTnc, which uses a combined DataEventStream. We can't
just add two mutable references to the same thing---Rust won't
allow it.

Refactor ArqState to take a combined sink/stream instead.