Fix panic on ArqStream drop
Previous versions of the ArqStream initiated a disconnect via the
default LocalPool. Unfortunately, the LocalPool cannot be used
from inside an async execution context, so this method was bound
to fail spectactularly.

Instead, spawn a thread, execute a future method that's part of
the ArdopTnc, and join it. This is ugly, but it works.