Switch to timer functions in crate runtime
Remove async-timer in favor of using the timeout() FuturesExt
method in crate runtime. We need runtime 0.3.0-alpha.6 or higher.
alpha.5 crashes spectacularly.

Also remove the threading behavior during disconnect-on-Drop. This
is not necessary… at least when crate runtime provides the
runtime. If it does, we can call the simple LocalPool-based
one directly.