Major changes to TncError
Add TncError::TimedOut to represent a generic, non-fatal
timeout.

Remove the TncError::CommandTimeout. These types of timeouts
should only occur if the TNC is dead. It is more appropriate
to represent them as TncError::IoError with a TimedOut error
kind.

Remove TncError::CommandResponseInvalid. This type of error
should only occur if the TNC is not operating as expected. It
is more appropriate to represent this as a TncError::IoError
instead.

The "event timeout" is also removed, to be replaced with more
specific timeouts.